[MOD] Refactor file structure
