+++ src/app/bundle.scss
@@ -1,5 +1,5 @@
-@import './variables';
-@import './global';
-@import './memberList/memberList';
-@import './addMember/addMember';
-@import './editMember/editMember';
+@import './styles/variables';
+@import './styles/global';
+@import './components/memberList/memberList';
+@import './components/addMember/addMember';
+@import './components/editMember/editMember';
