+++ src/app/styles/variables.scss
